% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2017-09-14.R
\name{as.data.frame.interval}
\alias{as.data.frame.interval}
\title{Transform interval objects into data.frames}
\usage{
\method{as.data.frame}{interval}(x, ...)
}
\arguments{
\item{x}{an object of class interval.}

\item{...}{arguments to be passed to \code{as.data.frame}.}
}
\value{
a data.frame containing x as a character
}
\description{
This function transforms interval objects into data.frames.
This is not only relevant on its own, it is also needed whenever
a function need objects as a data.frame (e.g. \code{View} or \code{cbind}).
}
