% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2017-12-28.R
\name{as.interval}
\alias{as.interval}
\title{Function to transform objects into an interval object}
\usage{
as.interval(x)
}
\arguments{
\item{x}{An object to transform.
Currently the function can transform numeric vectors and characters}
}
\value{
A vector of class \code{interval}.
}
\description{
Function to transform numeric (or character) vectors or n times 2 matrices into an interval object
}
\examples{
as.interval(c("1000;2000", "700:700"))
as.interval(c("1500;1500", "700:700"))
}
\seealso{
\link[hmi]{generate_interval}
}
