% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2017-04-20.R
\name{as_interval}
\alias{as_interval}
\title{Function to transform numeric (or character) vectors into an interval object}
\usage{
as_interval(lower, upper)
}
\arguments{
\item{lower}{a vector with the lower values of a variable}

\item{upper}{a vector with the upper values of a variable}
}
\value{
a character vector with the lower and upper bound values.
}
\description{
Function to transform numeric (or character) vectors into an interval object
}
