% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2016-07-05.R, R/hmi_smallfunctions_2016-07-14.R
\name{calc}
\alias{calc}
\title{helper function to be called in \code{update.alpha.coef}.}
\usage{
calc(x, sigma.y.new, sigma.alpha.new.inv)

calc(x, sigma.y.new, sigma.alpha.new.inv)
}
\arguments{
\item{x}{A data.frame which is basically cbind(Z_obs, y_obs - X_obs * beta.new).}

\item{sigma.y.new}{The numeric residual variance.}

\item{sigma.alpha.new.inv}{The inverse random effects covariance matrix.}

\item{x}{A data.frame which is basically cbind(Z_obs, y_obs - X_obs * beta.new).}

\item{sigma.y.new}{The numeric residual variance.}

\item{sigma.alpha.new.inv}{The inverse random effects covariance matrix.}
}
\value{
It returns (for this given cluster) a (cluster specific) random intercept and random slope.

It returns (for this given cluster) a (cluster specific) random intercept and random slope.
}
\description{
It (basically) calculates (Z'Z + sigma.e * sigma.b^-1)^-1 * z' resid.

It (basically) calculates (Z'Z + sigma.e * sigma.b^-1)^-1 * z' resid.
}

