% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions.R
\name{ceiling.interval}
\alias{ceiling.interval}
\title{Ceiling funtion for intervals}
\usage{
\method{ceiling}{interval}(x, ...)
}
\arguments{
\item{x}{numeric vector or interval object}

\item{...}{further arguments passed to \code{ceiling}.}
}
\description{
Ceiling funtion for intervals
}
