% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-05-07.R
\name{ceiling}
\alias{ceiling}
\title{Generic ceiling}
\usage{
ceiling(x, ...)
}
\arguments{
\item{x}{numeric vector or interval object}

\item{...}{further arguments passed to \code{ceiling}.}
}
\description{
Generic ceiling
}
