% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-05-07.R
\name{center.interval}
\alias{center.interval}
\title{Function to give the center of the interval}
\usage{
center.interval(interval, inf2NA = FALSE)
}
\arguments{
\item{interval}{an object from class interval}

\item{inf2NA}{logical. If \code{TRUE}, entries containing -Inf or Inf, will return NA.}
}
\value{
A numeric vector
}
\description{
Function to give the center of the interval object
}
