% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-05-07.R
\name{cleanup}
\alias{cleanup}
\title{cleanup data.frames}
\usage{
cleanup(X, k = Inf)
}
\arguments{
\item{X}{A n times p data.frame with p fixed (or random) effects variables.}

\item{k}{An integer defining the allowed maximum of levels in a factor covariate.
Currently (2018-03-13) not used.}
}
\value{
A n times (p-r) data.frame, with r being the number of variables with too many factors.
}
\description{
Function to exclude variables that have too many levels
(as they may cause numerical problems), to change binary factors to 0-1 coding
(as such factors might generate linear dependent variables) and
to remove multiple intercepts.
}
