% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2017-02-21.R
\name{contributions4intervalls}
\alias{contributions4intervalls}
\title{get the likelihood contributions of interval data}
\usage{
contributions4intervalls(lower, upper, mymean, mysd)
}
\arguments{
\item{lower}{a vector with the lower bounds of an interval covariate.}

\item{upper}{a vector with the upper bounds of an interval covariate.}

\item{mymean}{a numeric for the expected value of the normal distribution
(which is one of the parameters trying to be optimized
so that the likelihood becomes maximized)}

\item{mysd}{a numeric for the standard devitation of the normal distribution
(which is one of the parameters trying to be optimized
so that the likelihood becomes maximized)}
}
\value{
a vector giving the likelihood contributions of the interval data.
}
\description{
This function calculates the likelihood contributions of interval data
}
