% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-05-31.R
\name{contributions4intervals}
\alias{contributions4intervals}
\title{get the likelihood contributions of interval data}
\usage{
contributions4intervals(lower_bounds, upper_bounds, mymean, mysd)
}
\arguments{
\item{lower_bounds}{a vector with the lower bounds of an interval covariate.}

\item{upper_bounds}{a vector with the upper bounds of an interval covariate.}

\item{mymean}{a numeric for the expected value of the normal distribution
(which is one of the parameters trying to be optimized
so that the likelihood becomes maximized)}

\item{mysd}{a numeric for the standard deviation of the normal distribution
(which is one of the parameters trying to be optimized
so that the likelihood becomes maximized)}
}
\value{
a vector giving the likelihood contributions of the interval data.
}
\description{
This function calculates the likelihood contributions of interval data
}
