% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2017-02-21.R
\name{decompose_interval}
\alias{decompose_interval}
\title{decompose up intervals}
\usage{
decompose_interval(interval)
}
\arguments{
\item{interval}{an interval object of length n
(if it is something else, it is returned unchanged)}
}
\value{
a data.frame with the precise observations, the lower and the upper bounds
as covariates.
}
\description{
This function decomposes an interval object up into precise observations
(e.g. "1850.23;1850.23" into 1850.23),
imprecise observations (e.g. "1800;1900") and
missing observations ("-Inf;Inf" into NA)
}
