% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2017-02-21.R
\name{divided_interval}
\alias{divided_interval}
\title{Dividing function}
\usage{
divided_interval(interval, x)
}
\arguments{
\item{interval}{an object from class interval}

\item{x}{an single element or vector for division}
}
\value{
an interval object
}
\description{
Function to divide single elements or vectors (of correct dimension) to the interval object
}
