% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2017-01-05.R
\name{doubleintegral}
\alias{doubleintegral}
\title{Function to calculate double integrals}
\usage{
doubleintegral(lower_inner, upper_inner, lower_outer, upper_outer, cdf, ...)
}
\arguments{
\item{lower_inner}{The vector for the lower bound for the inner integral}

\item{upper_inner}{The vector for the upper bound for the inner integral}

\item{lower_outer}{The vector for the lower bound for the outer integral}

\item{upper_outer}{The vector for the upper bound for the outer integral}

\item{cdf}{the cumulative density function (from the class "function")}

\item{...}{Further arguments passed to the cdf.}
}
\value{
a vector with the value of the double integral for each observation (with an observed target variable)
}
\description{
This function is primarily build to make the funtion \code{components} more neat.
}

