% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions.R
\name{exp.interval}
\alias{exp.interval}
\title{Exp function for interval objects}
\usage{
\method{exp}{interval}(x, ...)
}
\arguments{
\item{x}{numeric vector or interval object}

\item{...}{further arguments passed to \code{exp}.}
}
\description{
Exp function for interval objects
}
