% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-05-31.R
\name{exp}
\alias{exp}
\alias{exp.default}
\alias{exp.interval}
\title{Generic exp}
\usage{
exp(x, ...)

\method{exp}{default}(x, ...)

\method{exp}{interval}(x, ...)
}
\arguments{
\item{x}{numeric vector or interval object}

\item{...}{further arguments passed to \code{exp}.}
}
\description{
Generic exp
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: for numeric vectors

\item \code{interval}: for interval objects.
}}

