% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions.R
\name{factors}
\alias{factors}
\title{Function to get all factors}
\usage{
factors(x)
}
\arguments{
\item{x}{A single integer; no vector.}
}
\value{
A numeric vector with the factors
}
\description{
Function to get all factors (not limited to prime factors) of an integer.
}
\references{
based on stackoverflow.com/questions/6424856 "R Function for returning ALL factors"
 answer by Chase
}
