% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-05-07.R
\name{fixed_intercept_check}
\alias{fixed_intercept_check}
\title{Function to check multilevel models on the existence of fixed intercepts}
\usage{
fixed_intercept_check(model_formula)
}
\arguments{
\item{model_formula}{A formula (from class \code{formula})}
}
\value{
A boolean value indicating whether there is a fixed intercept in the model or not
}
\description{
Function to check multilevel models on the existence of fixed intercepts.
The specification of an intercept by calling a 1-column (e.g. "int")
is not counted towards the existence of an intercept.
Contradictory inputs like "~ 1 + 0 + X1 + ..." or "~ -1 + 1 + X1 + ..."
will throw an error.
}
