% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions.R
\name{floor.interval}
\alias{floor.interval}
\title{Floor function for interval objects
Floor function for interval objects}
\usage{
\method{floor}{interval}(x, ...)
}
\arguments{
\item{x}{numeric vector or interval object}

\item{...}{further arguments passed to \code{floor}.}
}
\description{
Floor function for interval objects
Floor function for interval objects
}
