% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-05-07.R
\name{floor}
\alias{floor}
\alias{floor.default}
\alias{floor.interval}
\title{Generic floor}
\usage{
floor(x, ...)

\method{floor}{default}(x, ...)

\method{floor}{interval}(x, ...)
}
\arguments{
\item{x}{numeric vector or interval object}

\item{...}{further arguments passed to \code{floor}.}
}
\description{
Generic floor
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: for numeric vectors

\item \code{interval}: for interval objects.
}}

