% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gcsemv.R
\docType{data}
\name{Gcsemv}
\alias{Gcsemv}
\title{Exam results on the GCSE}
\format{A data frame with 1905 rows and 5 variables:
\describe{
\item{school}{The ID of the student's school.}
\item{student}{The ID of the student within its school.
They are not unique, different schools may have identical IDs for not identical students.}
\item{gender}{The gender of the student: 0 for boys and 1 for girls.}
\item{written}{The score in a written questionnaire.}
\item{coursework}{The score in a coursework.}
}}
\source{
Website of the Centre for Multilevel Modelling at the University of Bristol:
\url{http://www.bristol.ac.uk/cmm/media/team/hg/msm-3rd-ed/gcsemv.xls}
}
\usage{
Gcsemv
}
\description{
A dataset containing results on the General Certificate of Secondary Education (GCSE)
exams in Great Britain for 1905 students in 73 Schools.
}
\references{
Creswell M. (1991). A Multilevel bivariate model. In R. Prosser,
J. Rasbash, H. Goldstein (eds.), Data Analysis with ML3. Institute of Education, London.

Goldstein H. (2011) Multilevel Statistical Models. 4 edition. Wiley, Chichseter (UK),
ISBN 978-0-470-74865-7.
}
\keyword{datasets}
