% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-05-31.R
\name{generate_interval}
\alias{generate_interval}
\title{Function to generate an interval object}
\usage{
generate_interval(lower, upper)
}
\arguments{
\item{lower}{a vector with the lower values of a variable}

\item{upper}{a vector with the upper values of a variable}
}
\value{
a character vector with the lower and upper bound values.
}
\description{
Function to generate transform into an
\code{interval} object from numeric (or character) vectors.
}
