% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2016-07-05.R, R/hmi_smallfunctions_2016-07-14.R
\name{get_type}
\alias{get_type}
\title{Get the type of variables.}
\usage{
get_type(variable)

get_type(variable)
}
\arguments{
\item{variable}{A variable from your dataset.}

\item{variable}{A variable from your dataset.}
}
\value{
A character denoting the type of \code{variable}.

A character denoting the type of \code{variable}.
}
\description{
Function checks wether a variable is: ...

Function checks wether a variable is: ...
\itemize{
 \item continuous (just many numeric values),
 \item semicontinuous (many numeric values but more than 5% of the data are 0),
 \item binary (just two different values - like 0s and 1s or "m" and "f"),
 \item categorical (the variable is a factor or has more than 3 different values)
 \item orderd categorical (the categorical variable is ordered.)
}
}
\details{
continuous (just many numeric values),
 semicontinuous (many numeric values but more than 5% of the data are 0),
 binary (just two different values - like 0s and 1s or "m" and "f"),
 categorical (the variable is a factor or has more than 3 different values)
 orderd categorical (the categorical variable is ordered.)
}

