% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_mainfunctions_2016-07-05.R, R/hmi_mainfunctions_2016-07-13.R
\name{gibbs.coef}
\alias{gibbs.coef}
\title{Gibbs sampler for random coefficient model}
\usage{
gibbs.coef(y_gibbs, X_gibbs, Z_gibbs, clID, n.iter = 100, M = 10,
  n.chains = 3, burn.in = 1/3, max.iter = 5000)

gibbs.coef(y_gibbs, X_gibbs, Z_gibbs, clID, n.iter = 100, M = 10,
  n.chains = 3, burn.in = 1/3, max.iter = 5000)
}
\arguments{
\item{y_gibbs}{A vector or data.frame with \code{ncol = 1} containing the target variable with the missing values.}

\item{X_gibbs}{A data.frame containing the covariates influencing \code{y} via fixed effects.
If rows with missing values in \code{X} should also be imputed, put all your variables in a data.frame (or matrix)}

\item{Z_gibbs}{A data.frame containing the covariates influencing \code{y} via random effects}

\item{clID}{A factor (should come as data.frame or vector) containing the cluster IDs.}

\item{n.iter}{An integer defining the number of
iterations that should be run in each bunch of iterations.}

\item{M}{An integer defining the number of imputations that should be made.}

\item{n.chains}{An integer defining the number of Markov chains to be made.}

\item{burn.in}{A numeric between 0 and 1 defining the percentage of draws from the gibbs sampler
that should be discarded as burn in.}

\item{max.iter}{An integer defining the maximum number of
iterations that should be run in total.}

\item{y}{A vector or data.frame with \code{ncol = 1} containing the target variable with the missing values.}

\item{X}{A vector a data.frame containing the covariates influencing \code{y} via fixed effects.
If rows with missing values in \code{X} should also be imputed, put all your variables in a data.frame (or matrix)}

\item{Z}{A vector a data.frame containing the covariates influencing \code{y} via random effects}

\item{cl.id}{A factor (should come as data.frame or vector) containing the cluster IDs.}

\item{m}{An integer defining the number of imputations that should be made.}

\item{n.iter}{An integer defining the number of
iterations that should be run in each bunch of iterations.}

\item{max.iter}{An integer defining the maximum number of
iterations that should be run in total.}

\item{n.chains}{An integer defining the number of Markov chains to be made.}

\item{burn.in}{A numeric between 0 and 1 defining the percentage of draws from the gibbs sampler
that should be discarded as burn in.}
}
\value{
It returns a multidimensional vector with the Markov chains
containing the imputation parameters needed for \code{imp_multi}.

It returns a multidimensional vector with the Markov chains
containing the imputation parameters needed for \code{imp_multi}.
}
\description{
It generates the Markov chains of the imputation parameters by drawing from their conditional distributions
until convergence

It generates the Markov chains of the imputation parameters by drawing from their conditional distributions
until convergence
}

