% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions.R
\name{head.interval}
\alias{head.interval}
\title{Head for intervals}
\usage{
\method{head}{interval}(x, ...)
}
\arguments{
\item{x}{vector, matrix, table, data.frame or interval object}

\item{...}{further arguments passed to \code{head}.}
}
\description{
Head function for intervals returning the first elements of an \code{interval} object
}
