% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-05-07.R
\name{head}
\alias{head}
\alias{head.default}
\alias{head.interval}
\title{Generic head}
\usage{
head(x, ...)

\method{head}{default}(x, ...)

\method{head}{interval}(x, ...)
}
\arguments{
\item{x}{numeric vector or interval object}

\item{...}{further arguments passed to \code{head}.}
}
\description{
Generic head
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: for numeric vectors

\item \code{interval}: for interval objects.
}}

