% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2017-02-21.R
\name{idf2interval}
\alias{idf2interval}
\title{Transform interval data frames into data.frames with interval variables}
\usage{
idf2interval(idf)
}
\arguments{
\item{idf}{an interval data frame (idf-object).}
}
\value{
interval an \code{interval}.
}
\description{
This function is the path from the linLIR package (Wiencierz, 2012) to this hmi package.
}
