% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_imp_binary_single_2016-12-23.R
\name{imp_binary_single}
\alias{imp_binary_single}
\title{The function for imputation of binary variables.}
\usage{
imp_binary_single(y_imp_multi, X_imp_multi)
}
\arguments{
\item{y_imp_multi}{A Vector with the variable to impute.}

\item{X_imp_multi}{A data.frame with the fixed effects variables.}
}
\value{
A n x 1 matrix. The column is one set of imputed y-variables.
}
\description{
The function is called by the wrapper.
}

