% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_imp_cont_single_2016-12-22.R
\name{imp_cont_single}
\alias{imp_cont_single}
\title{The function for imputation of contiuous variables.}
\usage{
imp_cont_single(y_imp_multi, X_imp_multi)
}
\arguments{
\item{y_imp_multi}{A Vector with the variable to impute.}

\item{X_imp_multi}{A data.frame with the fixed effects variables.}
}
\value{
A n x 1 matrix.
}
\description{
The function is called by the wrapper (hmi). It uses \code{mice} with the method "norm".
}

