% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_imp_interval_2017-02-05.R
\name{imp_interval}
\alias{imp_interval}
\title{The function to impute interval data variables}
\usage{
imp_interval(y_imp_multi, X_imp_multi, M)
}
\arguments{
\item{y_imp_multi}{A Vector with the variable to impute.}

\item{X_imp_multi}{A data.frame with the fixed effects variables.}

\item{M}{An integer defining the number of imputations that should be made.}
}
\value{
A n x M matrix. Each column is one of M imputed y-variables.
}
\description{
This functions imputes interval data variables. Those are variables,
that consists of a lower and upper (numeric) boundary. Technically
those boundaries are cointained in a string, seperated by a semi colon.
E.g. if a person reports there income to be something between 3000 and 4000 dollars,
its value in the interval covariate would be \code{"3000;4000"}.
Left (resp. right) censored data can be donoted by \code{"-Inf;x"} (resp. \code{"x;Inf"}),
with \code{x} being the (numeric) observed value.
}
