% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_imp_catordered_multi_2017-01-04.R
\name{imp_orderedcat_multi}
\alias{imp_orderedcat_multi}
\title{The function for hierarchical imputation of categorical variables.}
\usage{
imp_orderedcat_multi(y_imp_multi, X_imp_multi, Z_imp_multi, clID, model_formula,
  M = 10, nitt = 3000, thin = 10, burnin = 1000)
}
\arguments{
\item{y_imp_multi}{A Vector with the variable to impute.}

\item{X_imp_multi}{A data.frame with the fixed effects variables.}

\item{Z_imp_multi}{A data.frame with the random effects variables.}

\item{clID}{A vector with the cluster ID.}

\item{model_formula}{A \code{\link[stats]{formula}} used for the analysis model.}

\item{M}{An integer defining the number of imputations that should be made.}

\item{nitt}{An integer defining number of MCMC iterations (see MCMCglmm).}

\item{thin}{An integer defining the thinning interval (see MCMCglmm).}

\item{burnin}{An integer defining the percentage of draws from the gibbs sampler
that should be discarded as burn in (see MCMCglmm).}
}
\value{
A n x M matrix. Each column is one of M imputed y-variables.
}
\description{
The function is called by the wrapper.
}

