% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_imp_catordered_single_2016-12-22.R
\name{imp_orderedcat_single}
\alias{imp_orderedcat_single}
\title{The function to impute ordered categorical variables}
\usage{
imp_orderedcat_single(y_imp_multi, X_imp_multi)
}
\arguments{
\item{y_imp_multi}{A Vector with the variable to impute.}

\item{X_imp_multi}{A data.frame with the fixed effects variables.}
}
\value{
A n x 1 matrix.
}
\description{
The function uses the proportional odds logistic regression (polr) approach,
implemented in \code{mice}.
}

