% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_imp_semicont_single_2016-12-10.R
\name{imp_semicont_single}
\alias{imp_semicont_single}
\title{The function for hierarchical imputation of semicontinous variables.}
\usage{
imp_semicont_single(y_imp_multi, X_imp_multi, heap = 0, M = 10)
}
\arguments{
\item{y_imp_multi}{A Vector with the variable to impute.}

\item{X_imp_multi}{A data.frame with the fixed effects variables.}

\item{heap}{A scalar saying to which value the data might be heaped.}

\item{M}{An integer defining the number of imputations that should be made.}
}
\value{
A n x M matrix. Each column is one of M imputed y-variables.
}
\description{
The function is called by the wrapper. We consider data to be "semicontinuous" when
more than 5\% of the (non categorical) observations.\cr
For example in surveys a certain portion of people, when asked for their income,
report "0", which clearly violates the assumption of income to be (log-) normally distributed.
}

