% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_inputcheck_2016-07-05.R
\name{inputcheck}
\alias{inputcheck}
\title{Input checker}
\usage{
inputcheck(data, model_formula, y, X, Z, cl.id, n.iter, M, n.chains, burn.in)
}
\arguments{
\item{data}{A matrix or (better) a data.frame with all variables appearing in \code{model_formula}.}

\item{model_formula}{A \code{\link[stats]{formula}} used for the analysis model.}

\item{y}{A vector or data.frame with \code{ncol = 1} containing the target variable with the missing values.}

\item{X}{A vector a data.frame containing the covariates influencing \code{y} via fixed effects.
If rows with missing values in \code{X} should also be imputed, put all your variables in a data.frame (or matrix)}

\item{Z}{A vector a data.frame containing the covariates influencing \code{y} via random effects}

\item{cl.id}{A factor (should come as data.frame or vector) containing the cluster IDs.}

\item{n.iter}{An integer defining the number of
iterations that should be run in each bunch of iterations.}

\item{M}{An integer defining the number of imputations that should be made.}

\item{n.chains}{An integer defining the number of Markov chains to be made.}

\item{burn.in}{A numeric between 0 and 1 defining the percentage of draws from the gibbs sampler
that should be discarded as burn in.}
}
\value{
Nothing. In some cases prompts might be cast.
}
\description{
A help function to check whether in input is correct.
}

