% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2017-09-14.R
\name{+.interval}
\alias{+.interval}
\title{Adding function}
\usage{
\method{+}{interval}(interval, x)
}
\arguments{
\item{interval}{an object from class interval}

\item{x}{an single element or vector to add to the interval object}
}
\value{
an interval object
}
\description{
Function to add single elements or vectors (of correct dimension) to the interval object
}
