% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-05-31.R
\name{/.interval}
\alias{/.interval}
\title{Dividing function}
\usage{
\method{/}{interval}(interval, x)
}
\arguments{
\item{interval}{an object from class interval}

\item{x}{an single element or vector for division}
}
\value{
an interval object
}
\description{
Function to divide single elements or vectors (of correct dimension) to the interval object
}
