% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions.R
\name{\%\%.interval}
\alias{\%\%.interval}
\title{Modulo function}
\usage{
\method{\%\%}{interval}(x, interval)
}
\arguments{
\item{x}{an single element or vector to add to the interval object}

\item{interval}{an object from class interval}
}
\value{
a vector with the modulo for the precise elements in \code{interval}.
For imprecise elements, \code{NA} is returned.
}
\description{
Modulo function for interval objects
}
