% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-05-07.R
\name{-.interval}
\alias{-.interval}
\title{Subtraction function}
\usage{
\method{-}{interval}(interval, x)
}
\arguments{
\item{interval}{an object from class interval}

\item{x}{an single element or vector to subtract to the interval object}
}
\value{
an interval object
}
\description{
Function to subtract single elements or vectors (of correct dimension) from the interval object
}
