% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-01-31.R
\name{is.na.interval}
\alias{is.na.interval}
\title{is.na for interval objects}
\usage{
\method{is.na}{interval}(interval)
}
\arguments{
\item{interval}{An \code{interval} object of length n}
}
\value{
A boolean vector of length n indicating whether the entries in \code{interval}
are \code{NA} or not. Cf. \code{is.na}.
}
\description{
This functions checks whether elements from an \code{interval} object are \code{NA}
}
