% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2017-09-14.R
\name{is_interval}
\alias{is_interval}
\title{Function to check whether an object is an interval}
\usage{
is_interval(x)
}
\arguments{
\item{x}{an object}
}
\value{
a boolean value indicting whether x is an interval or not
}
\description{
If there are numerics separated by semicolons (;), this is considered to be an interval.
intervals with 2.4e5 are not considered to be an interval.
}
