% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions.R
\name{list_of_rounding_degrees_maker}
\alias{list_of_rounding_degrees_maker}
\title{Helps the user to make a list of rounding degrees}
\usage{
list_of_rounding_degrees_maker(data)
}
\arguments{
\item{data}{the data.frame also passed to \code{hmi}.}
}
\value{
a list with suggested rounding degrees. Each list element has the name
of a rounded continuous variable in the data.frame. The elements contain a
numeric vector with the rounding degrees found for that variable.
}
\description{
In \code{hmi} the user can add a list of rounding degrees. This function gives him a framework
with suggestions. Of course the user can make changes by herself/himself afterwards.
For example, the function might wrongly classify a variable to be heaped or selects
unwanted rounding degrees.
}
