% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-05-07.R
\name{list_of_rounding_formulas_maker}
\alias{list_of_rounding_formulas_maker}
\title{Helps the user to make a list of rounding formulas for the rounding degrees}
\usage{
list_of_rounding_formulas_maker(data, default = ~.)
}
\arguments{
\item{data}{the data.frame also passed to \code{hmi}.}

\item{default}{A default formula used for every rounded variable.}
}
\value{
a list with suggested rounding degree formulas. Each list element has the name
of a rounded continuous variable in the data.frame. The elements contain a
very general rounding degree formula.
}
\description{
In \code{hmi} the user can add a list of rounding formulas for each variable suffering rounding.
This function gives him/her a framework with suggestions. Of course the user can make changes
by herself/himself afterwards.
For example, the function might wrongly classify a variable to be heaped.
}
