% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions.R
\name{list_of_spikes_maker}
\alias{list_of_spikes_maker}
\title{Helps the user to make a list of spikes.}
\usage{
list_of_spikes_maker(data)
}
\arguments{
\item{data}{the data.frame also passed to \code{hmi}.}
}
\value{
a list with suggested spikes. Each list element has the name of a spiked variable
in the data.frame. The elements contain a single numeric denoting the spike found for that variable.
}
\description{
In \code{hmi} the user can add a list of spikes. This function gives her/him a framework
with suggestions. Of course the user can make changes by herself/himself afterwards.
For example, the function might wrongly classify a variable to have a spike.
}
