% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions.R
\name{log.interval}
\alias{log.interval}
\title{Log function for interval objects}
\usage{
\method{log}{interval}(x, ...)
}
\arguments{
\item{x}{numeric vector or interval object}

\item{...}{further arguments passed to \code{log}}
}
\description{
Log function for interval objects
}
