% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions.R
\name{log}
\alias{log}
\alias{log.default}
\alias{log.interval}
\title{Generic log}
\usage{
log(x, ...)

\method{log}{default}(x, ...)

\method{log}{interval}(x, ...)
}
\arguments{
\item{x}{numeric vector or interval object}

\item{...}{further arguments passed to \code{log}}
}
\description{
Generic log
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: for numeric vectors

\item \code{interval}: for interval objects.
}}

