% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions.R
\name{Mode}
\alias{Mode}
\title{Get the mode}
\usage{
Mode(x)
}
\arguments{
\item{x}{A vector}
}
\value{
The mode of x as a numeric value.
}
\description{
This function calculates the mode (most frequent observation) of a vector.
}
\references{
Adopted from stackoverflow.com/questions/2547402: "is there a built in function for finding the mode"
from user "Ken Williams".
}
