% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2017-09-14.R
\name{negloglik2_intervalsonly}
\alias{negloglik2_intervalsonly}
\title{calculate the likelihood contribution of interval data only}
\usage{
negloglik2_intervalsonly(para, X, lower, upper)
}
\arguments{
\item{para}{This is the vector \eqn{Psi} of parameters defining model
(see p. 62 in Drechsler, Kiesl & Speidel, 2015).
With respect to them, the value returned by this function shall be
maximized.\cr
The starting values are c(betastart2, sigmastart2)
(the regression parameters explaining the logged income,
and the variance parameter).}

\item{X}{the data.frame of covariates.}

\item{lower}{the lower bound of an interval variable.}

\item{upper}{the upper bound of an interval variable.}
}
\value{
An integer equal to the (sum of the) negative log-likelihood contributions (of the observations)
}
\description{
calculate the likelihood contribution of interval data only
}
\references{
Joerg Drechsler, Hans Kiesl, Matthias Speidel (2015):
"MI Double Feature: Multiple Imputation to Address Nonresponse and Rounding Errors in Income Questions",
Austrian Journal of Statistics, Vol. 44, No. 2, \url{http://dx.doi.org/10.17713/ajs.v44i2.77}
}
