% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2017-01-05.R
\name{pbivnormX}
\alias{pbivnormX}
\title{calculate probabilities from the cumulative distribution function of a standard bivariate normal distribution}
\usage{
pbivnormX(x, y, rho = 0)
}
\arguments{
\item{x}{the vector (or a two columned matrix) with the values of the first random variable}

\item{y}{the vector with the values of the second random variable}

\item{rho}{the correlation (a scalar) between the two random variables.}
}
\value{
A vector with the values of the density distribution at the points (x, y).
}
\description{
A modfied version of pbivnorm() from package \code{pbivnorm}.
It is needed in the imputation routine for rounded income.
}

