% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2017-09-14.R
\name{remove_excessives}
\alias{remove_excessives}
\title{Removing excessive factors}
\usage{
remove_excessives(X, k = 10)
}
\arguments{
\item{X}{A n times p data.frame with p fixed (or random) effects variables.}

\item{k}{An integer defining the allowed maximum of levels in a factor covariate.}
}
\value{
A n times (p-r) data.frame, with r being the number of variables with too many factors.
}
\description{
Function to exclude variables that have too many levels as they may cause numerical problems
}
