% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2017-11-14.R
\name{resetPar}
\alias{resetPar}
\title{Function to reset all graphics parameters}
\usage{
resetPar()
}
\description{
With this function other functions can manipulate the \code{par} settings and
when they are finished restore the settings to the state before the function call.
The package \code{hmi} uses this mechanism for its function \code{chaincheck}.
}
\references{
https://stackoverflow.com/questions/5789982/reset-par-to-the-default-values-at-startup
}
\author{
Gavin Simpson
}
