% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions.R
\name{round.interval}
\alias{round.interval}
\title{Round function for interval objects}
\usage{
\method{round}{interval}(x, ...)
}
\arguments{
\item{x}{numeric vector or interval object}

\item{...}{further arguments passed to \code{round}.}
}
\description{
Round function for interval objects
}
