% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions.R
\name{round}
\alias{round}
\alias{round.default}
\alias{round.interval}
\title{Generic round}
\usage{
round(x, ...)

\method{round}{default}(x, ...)

\method{round}{interval}(x, ...)
}
\arguments{
\item{x}{numeric vector or interval object}

\item{...}{further arguments passed to \code{round}.}
}
\description{
Generic round
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: for numeric vectors

\item \code{interval}: for interval objects.
}}

