% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2016-07-05.R, R/hmi_smallfunctions_2016-07-14.R
\name{sample_imp}
\alias{sample_imp}
\title{Sample imputation.}
\usage{
sample_imp(variable)

sample_imp(variable)
}
\arguments{
\item{variable}{A vector of size \code{n} with missing values.}

\item{variable}{A vector of size \code{n} with missing values.}
}
\value{
A vector of size \code{n} without missing values.

A vector of size \code{n} without missing values.
}
\description{
Function to sample values in a variable from other (observed) values in this variable.
So this imputation doesn't use further covariates.

Function to sample values in a variable from other (observed) values in this variable.
So this imputation doesn't use further covariates.
}
\examples{
sample_imp(c(1, NA, 3, NA, 5))
sample_imp(c(1, NA, 3, NA, 5))
}

