% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions.R
\name{sna_interval}
\alias{sna_interval}
\title{Get standard NAs from interval data}
\usage{
sna_interval(x)
}
\arguments{
\item{x}{can by any object, but the function was designed for \code{interval}-objects.}
}
\value{
In case of \code{x} being an \code{interval}-object, it returns a n times 2 matrix.
 The first column is the lower bound, the second the upper bound.
 Otherwise it returns just \code{x}.
}
\description{
This function replaces observations with "-Inf;Inf" in an interval,
which basically means "no information available", with the standard NAs
(therefore the name 'sna').
Observations with a finite bound (e.g.x = "0;Inf") are not replaced,
since they contain information (here: "x is positive").
}
