% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-01-31.R
\name{split_interval}
\alias{split_interval}
\title{Split up intervals}
\usage{
split_interval(interval)
}
\arguments{
\item{interval}{an interval object of length n
(if it is something else, it is returned unchanged)}
}
\value{
a n times 2 matrix. The first column is the lower bound, the second the upper bound.
}
\description{
This function splits an interval object up into the lower and upper bound
}
