% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions.R
\name{sqrt.interval}
\alias{sqrt.interval}
\title{Sqrt function for interval objects}
\usage{
\method{sqrt}{interval}(x, ...)
}
\arguments{
\item{x}{numeric vector or interval object}

\item{...}{further arguments passed to \code{sqrt}.}
}
\description{
Sqrt function for interval objects
}
