% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-05-31.R
\name{sqrt}
\alias{sqrt}
\alias{sqrt.default}
\alias{sqrt.interval}
\title{Generic sqrt}
\usage{
sqrt(x, ...)

\method{sqrt}{default}(x, ...)

\method{sqrt}{interval}(x, ...)
}
\arguments{
\item{x}{numeric vector or interval object}

\item{...}{further arguments passed to \code{sqrt}.}
}
\description{
Generic sqrt
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: for numeric vectors

\item \code{interval}: for interval objects.
}}

