% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-05-07.R
\name{stand}
\alias{stand}
\title{Standardizing function}
\usage{
stand(X)
}
\arguments{
\item{X}{A n times p data.frame with p fixed (or random) effects variables.}
}
\value{
A n times p data.frame with the standardized versions of the numeric variables.
}
\description{
Function to standardize variables that are numeric (continuous and count variables) but no rounded continuous, semicontinuous, intercepts or categorical variables.
}
