% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions.R
\name{[.interval}
\alias{[.interval}
\title{Index for interval}
\usage{
\method{[}{interval}(obj, index)
}
\arguments{
\item{obj}{the interval object}

\item{index}{the index of the elements to replace}
}
\description{
Function to index elements of an interval object
}
