% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-01-31.R
\name{suggest_rounding_degrees}
\alias{suggest_rounding_degrees}
\title{suggesting rounding degrees}
\usage{
suggest_rounding_degrees(x)
}
\arguments{
\item{x}{A vector or \code{interval} object.}
}
\description{
A function that suggests some rounding degrees of a continuous variable
(classically formatted or as interval object)
}
