% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-02-28.R
\name{table}
\alias{table}
\alias{table.interval}
\alias{table.default}
\title{Tabulating interval objects}
\usage{
table(x, ...)

\method{table}{interval}(x, ...)

\method{table}{default}(x, ...)
}
\arguments{
\item{x}{In its most save way, \code{x} is an object from class \code{interval}.}

\item{...}{Other parameters passed to \code{table}.}
}
\value{
A table.
}
\description{
Function to tabulate interval objects
}
