% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-01-31.R
\name{table_interval}
\alias{table_interval}
\title{Tabulating only interval objects}
\usage{
table_interval(x, ...)
}
\arguments{
\item{x}{In its most save way, \code{x} is an object from class \code{interval}.}

\item{...}{Other parameters passed to \code{table}.}
}
\value{
A table.
}
\description{
Function to only tabulate interval objects
}
