% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions.R
\name{tail.interval}
\alias{tail.interval}
\title{Tail for intervals}
\usage{
\method{tail}{interval}(x, ...)
}
\arguments{
\item{x}{vector, matrix, table, data.frame or interval object}

\item{...}{further arguments passed to \code{tail}.}
}
\description{
Tail function for intervals returning the last elements of an \code{interval} object
}
