% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-05-31.R
\name{tail}
\alias{tail}
\alias{tail.default}
\alias{tail.interval}
\title{Generic tail}
\usage{
tail(x, ...)

\method{tail}{default}(x, ...)

\method{tail}{interval}(x, ...)
}
\arguments{
\item{x}{numeric vector or interval object}

\item{...}{further arguments passed to \code{tail}.}
}
\description{
Generic tail
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: for numeric vectors

\item \code{interval}: for interval objects.
}}

