% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2016-07-05.R, R/hmi_smallfunctions_2016-07-14.R
\name{update.alpha.coef}
\alias{update.alpha.coef}
\title{update.alpha.coef}
\usage{
\method{update}{alpha.coef}(Z_obs, y_obs, X_obs, beta.new, clID_obs,
  sigma.y.new, sigma.alpha.new.inv)

\method{update}{alpha.coef}(Z_obs, y_obs, X_obs, beta.new, clID_obs,
  sigma.y.new, sigma.alpha.new.inv)
}
\arguments{
\item{Z_obs}{The random effects data matrix of those observations with an observed value of y.}

\item{y_obs}{The target variable of those observations with an observed value of y.}

\item{X_obs}{The fixed effects data matrix of those observations with an observed value of y.}

\item{beta.new}{The vector of fixed effects parameters.}

\item{clID_obs}{The cluster ID vector of those observations with an observed value of y.}

\item{sigma.y.new}{The numeric residual variance.}

\item{sigma.alpha.new.inv}{The inverse random effects covariance matrix.}

\item{Z_obs}{The random effects data matrix of those observations with an observed value of y.}

\item{y_obs}{The target variable of those observations with an observed value of y.}

\item{X_obs}{The fixed effects data matrix of those observations with an observed value of y.}

\item{beta.new}{The vector of fixed effects parameters.}

\item{clID_obs}{The cluster ID vector of those observations with an observed value of y.}

\item{sigma.y.new}{The numeric residual variance.}

\item{sigma.alpha.new.inv}{The inverse random effects covariance matrix.}
}
\value{
It returns cluster specific random effects.

It returns cluster specific random effects.
}
\description{
The function updates the cluster specific random effects
by drawing from a multivariate normal distribution
(with a unique distribution for every cluster).

The function updates the cluster specific random effects
by drawing from a multivariate normal distribution
(with a unique distribution for every cluster).
}

