% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2016-07-05.R, R/hmi_smallfunctions_2016-07-14.R
\name{update.beta.coef}
\alias{update.beta.coef}
\title{update.beta.coef}
\usage{
\method{update}{beta.coef}(Z_obs, alpha.new, b.hat.part1, clID_obs, y_obs,
  sigma.y.new, xtx)

\method{update}{beta.coef}(Z_obs, alpha.new, b.hat.part1, clID_obs, y_obs,
  sigma.y.new, xtx)
}
\arguments{
\item{Z_obs}{A data.frame with the random effects variables.}

\item{alpha.new}{A matrix (with the dimension: number of clusters times number of random effect variables)
of cluster specific random effects.}

\item{b.hat.part1}{The (X'X)^{-1} X' matrix.}

\item{clID_obs}{The cluster ID vector of those observations with an observed value of y.}

\item{y_obs}{The target variable of those observations with an observed value of y.}

\item{sigma.y.new}{The residual variance (positive numeric).}

\item{xtx}{The (X'X)^{-1} matrix.}

\item{Z_obs}{A data.frame with the random effects variables.}

\item{alpha.new}{A matrix (with the dimension: number of clusters times number of random effect variables)
of cluster specific random effects.}

\item{b.hat.part1}{The (X'X)^{-1} X' matrix.}

\item{clID_obs}{The cluster ID vector of those observations with an observed value of y.}

\item{y_obs}{The target variable of those observations with an observed value of y.}

\item{sigma.y.new}{The residual variance (positive numeric).}

\item{xtx}{The (X'X)^{-1} matrix.}
}
\value{
It returns a vector of new fixed effects parameters.

It returns a vector of new fixed effects parameters.
}
\description{
The functions updates the fixed effects (imputation) parameters.
By drawing them from their normal distribution.

The functions updates the fixed effects (imputation) parameters.
By drawing them from their normal distribution.
}

