% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2016-07-05.R, R/hmi_smallfunctions_2016-07-14.R
\name{update.sigma.alpha.coef}
\alias{update.sigma.alpha.coef}
\title{update.sigma.alpha.coef}
\usage{
\method{update}{sigma.alpha.coef}(hyper.df1, hyper.df2, alpha.new)

\method{update}{sigma.alpha.coef}(hyper.df1, hyper.df2, alpha.new)
}
\arguments{
\item{hyper.df1}{A vector of numeric hyper parameters}

\item{hyper.df2}{A matrix of hyper parameters.}

\item{alpha.new}{A matrix of cluster specific random effects.}

\item{hyper.df1}{A vector of numeric hyper parameters}

\item{hyper.df2}{A matrix of hyper parameters.}

\item{alpha.new}{A matrix of cluster specific random effects.}
}
\value{
It returns a matrix with a new random effects covarariance matrix.

It returns a matrix with a new random effects covarariance matrix.
}
\description{
It updates the random effects covarariance matrix by drawing from a Wishart distribution.

It updates the random effects covarariance matrix by drawing from a Wishart distribution.
}

