% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_wrapper_2016-07-11.R, R/hmi_wrapper_2016-07-12.R, R/hmi_wrapper_2016-07-13.R
\name{wrapper}
\alias{wrapper}
\title{The main warpper function called by the user. LATER THE USER WILL ONLY USE THE WRAPPER.}
\usage{
wrapper(data, model_formula = NULL, parameter_list = NULL, n.iter = 100,
  M = 10, n.chains = 3, burn.in = 1/3, max.iter = 5000)

wrapper(data, model_formula = NULL, parameter_list = NULL, n.iter = 100,
  M = 10, n.chains = 3, burn.in = 1/3, max.iter = 5000)

wrapper(data, model_formula = NULL, parameter_list = NULL, n.iter = 100,
  M = 10, n.chains = 3, burn.in = 1/3, max.iter = 5000)
}
\arguments{
\item{data}{A matrix or (better) a data.frame with all variables appearing in \code{model_formula}.}

\item{model_formula}{A \code{\link[stats]{formula}} used for the analysis model.}

\item{parameter_list}{A LIST OF PARAMETERS}

\item{n.iter}{An integer defining the number of
iterations that should be run in each bunch of iterations.}

\item{M}{An integer defining the number of imputations that should be made.}

\item{n.chains}{An integer defining the number of Markov chains to be made.}

\item{burn.in}{A numeric between 0 and 1 defining the percentage of draws from the gibbs sampler
that should be discarded as burn in.}

\item{max.iter}{An integer defining the maximum number of
iterations that should be run in total.}

\item{impsyn}{SHALL VALUES BE IMPUTED OR SYNTHETICISED?}

\item{parameter.list}{A LIST OF PARAMETERS}

\item{data}{A matrix or (better) a data.frame with all variables appearing in \code{model_formula}.}

\item{model_formula}{A \code{\link[stats]{formula}} used for the analysis model.}

\item{M}{An integer defining the number of imputations that should be made.}

\item{data}{A matrix or (better) a data.frame with all variables appearing in \code{model_formula}.}

\item{model_formula}{A \code{\link[stats]{formula}} used for the analysis model.}

\item{parameter_list}{A LIST OF PARAMETERS}

\item{n.iter}{An integer defining the number of
iterations that should be run in each bunch of iterations.}

\item{M}{An integer defining the number of imputations that should be made.}

\item{n.chains}{An integer defining the number of Markov chains to be made.}

\item{burn.in}{A numeric between 0 and 1 defining the percentage of draws from the gibbs sampler
that should be discarded as burn in.}
}
\value{
A data.frame. It consists of the original \code{data} and \code{m}
additional variables with the imputed values.

A data.frame. It consists of the original \code{data} and \code{m}
additional variables with the imputed values.

A data.frame. It consists of the original \code{data} and \code{m}
additional variables with the imputed values.
}
\description{
The user has to passes to the function his data.
Optionally he pass his analysis model formula so that hmi runs the imputation model
in line with his analysis model formula.
And of course he can specify some parameters for the imputation routine
(like the number of imputations and iterations and the burn in percentage.)
The standard usage should be that the user gives his complete dataset
and his analysis model. But he also could just give y, X and Z and the cluser ID.

The user has to passes to the function his data.
Optionally he pass his analysis model formula so that hmi runs the imputation model
in line with his analysis model formula.
And of course he can specify some parameters for the imputation routine
(like the number of imputations and iterations and the burn in percentage.)
The standard usage should be that the user gives his complete dataset
and his analysis model. But he also could just give y, X and Z and the cluser ID.

The user has to passes to the function his data.
Optionally he pass his analysis model formula so that hmi runs the imputation model
in line with his analysis model formula.
And of course he can specify some parameters for the imputation routine
(like the number of imputations and iterations and the burn in percentage.)
The standard usage should be that the user gives his complete dataset
and his analysis model. But he also could just give y, X and Z and the cluser ID.
}

