% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-05-31.R
\name{ceiling}
\alias{ceiling}
\alias{ceiling.default}
\alias{ceiling.interval}
\title{Generic ceiling}
\usage{
ceiling(x, ...)

\method{ceiling}{default}(x, ...)

\method{ceiling}{interval}(x, ...)
}
\arguments{
\item{x}{numeric vector or interval object}

\item{...}{further arguments passed to \code{ceiling}.}
}
\description{
Generic ceiling
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: for numeric vectors

\item \code{interval}: for interval objects.
}}

