% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions.R
\name{get_type}
\alias{get_type}
\title{Get the type of variables.}
\usage{
get_type(variable, spike = NULL, rounding_degrees = c(1, 10, 100, 1000))
}
\arguments{
\item{variable}{A variable (vector) from your data set.}

\item{spike}{A numeric value, denoting the presumed spike of semi-continuous variables.}

\item{rounding_degrees}{A numeric vector with the presumed rounding degrees.
If the rounding_degrees are set to be \code{NULL}, the proposed rounding degrees from the function
\code{suggest_rounding_degrees} are used.}
}
\value{
A character denoting the type of \code{variable}.
}
\description{
Function checks of which type a variable is. The types are listed below
(together with a rough summary of the classification mechanism).
\itemize{
 \item continuous (numeric values, or integers with more than 20 different values),
 \item semicontinuous (numeric values with more than 10\% of them share the same value),
 \item rounded continuous (if more than 50\% of the observations of a continuous variable
  are divisible by some rounding degrees)
 \item count data (integers).
 \item an intercept (the same value for all observations),
 \item binary (two different values - like 0s and 1s or "m" and "f"),
 \item categorical (the variable is a factor or has more than 3 different values)
 \item ordered categorical (the categorical variable is ordered.)
}
}
\examples{
get_type(iris$Sepal.Length); get_type(iris$Species)
}
