% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-02-28.R
\name{idf2interval}
\alias{idf2interval}
\title{Transform interval data frames into data.frames with interval variables}
\usage{
idf2interval(idf)
}
\arguments{
\item{idf}{an interval data frame (idf-object).}
}
\value{
A \code{data.frame} where the interval variables are stored as \code{interval} objects.
}
\description{
This function is the path from the linLIR package (Wiencierz, 2012) to this hmi package.
}
