% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_imp_binary_single_2017-04-11.R
\name{imp_binary_single}
\alias{imp_binary_single}
\title{The function for imputation of binary variables.}
\usage{
imp_binary_single(y_imp, X_imp, rounding_degrees = c(1, 10, 100, 1000))
}
\arguments{
\item{y_imp}{A Vector with the variable to impute.}

\item{X_imp}{A data.frame with the fixed effects variables.}

\item{rounding_degrees}{A numeric vector with the presumed rounding degrees.}
}
\value{
A n x 1 data.frame with the original and imputed values.
}
\description{
The function is called by the wrapper.
}
