% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_imp_binary_single.R
\name{imp_binary_single}
\alias{imp_binary_single}
\title{The function for imputation of binary variables.}
\usage{
imp_binary_single(y_imp, X_imp, pvalue = 0.2, k = Inf)
}
\arguments{
\item{y_imp}{A Vector with the variable to impute.}

\item{X_imp}{A data.frame with the fixed effects variables.}

\item{pvalue}{A numeric between 0 and 1 denoting the threshold of p-values a variable in the imputation
model should not exceed. If they do, they are excluded from the imputation model.}

\item{k}{An integer defining the allowed maximum of levels in a factor covariate.}
}
\value{
A n x 1 data.frame with the original and imputed values.
}
\description{
The function is called by the wrapper.
}
