% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_imp_cont_multi_2017-04-11.R
\name{imp_cont_multi}
\alias{imp_cont_multi}
\title{The function for hierarchical imputation of contious variables.}
\usage{
imp_cont_multi(y_imp, X_imp, Z_imp, clID, nitt = 3000, thin = 10,
  burnin = 1000)
}
\arguments{
\item{y_imp}{A Vector with the variable to impute.}

\item{X_imp}{A data.frame with the fixed effects variables.}

\item{Z_imp}{A data.frame with the random effects variables.}

\item{clID}{A vector with the cluster ID.}

\item{nitt}{An integer defining number of MCMC iterations (see MCMCglmm).}

\item{thin}{An integer defining the thinning interval (see MCMCglmm).}

\item{burnin}{An integer defining the percentage of draws from the gibbs sampler
that should be discarded as burn in (see MCMCglmm).}
}
\value{
A n x 1 matrix with the original and imputed values.
}
\description{
The function is called by the wrapper.
}
