% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_imp_cont_multi_2018-04-17.R
\name{imp_cont_multi}
\alias{imp_cont_multi}
\title{The function for hierarchical imputation of continuous variables.}
\usage{
imp_cont_multi(y_imp, X_imp, Z_imp, clID, nitt = 22000, burnin = 2000,
  thin = 20, pvalue = 0.2, k = Inf)
}
\arguments{
\item{y_imp}{A vector with the variable to impute.}

\item{X_imp}{A data.frame with the fixed effects variables.}

\item{Z_imp}{A data.frame with the random effects variables.}

\item{clID}{A vector with the cluster ID.}

\item{nitt}{An integer defining number of MCMC iterations (see MCMCglmm).}

\item{burnin}{burnin A numeric value between 0 and 1 for the desired percentage of
Gibbs samples that shall be regarded as burnin.}

\item{thin}{An integer to set the thinning interval range. If thin = 1,
every iteration of the Gibbs-sampling chain will be kept. For highly autocorrelated
chains, that are only examined by few iterations (say less than 1000).}

\item{pvalue}{A numeric between 0 and 1 denoting the threshold of p-values a variable in the imputation
model should not exceed. If they do, they are excluded from the imputation model.}

\item{k}{An integer defining the allowed maximum of levels in a factor covariate.}
}
\value{
A list with 1. 'y_ret' the n x 1 data.frame with the original and imputed values.
2. 'Sol' the Gibbs-samples for the fixed effects parameters.
3. 'VCV' the Gibbs-samples for variance parameters.
}
\description{
The function is called by the wrapper.
}
