% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_imp_cont_single_2017-04-11.R
\name{imp_cont_single}
\alias{imp_cont_single}
\title{The function for imputation of continuous variables.}
\usage{
imp_cont_single(y_imp, X_imp)
}
\arguments{
\item{y_imp}{A Vector with the variable to impute.}

\item{X_imp}{A data.frame with the fixed effects variables.}
}
\value{
A n x 1 data.frame with the original and imputed values.
}
\description{
The function is called by the wrapper (hmi). It uses \code{mice} with the method "norm".
}
