% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_imp_catordered_single_2017-08-02.R
\name{imp_orderedcat_single}
\alias{imp_orderedcat_single}
\title{The function to impute ordered categorical variables}
\usage{
imp_orderedcat_single(y_imp, X_imp, rounding_degrees = c(1, 10, 100, 1000))
}
\arguments{
\item{y_imp}{A Vector with the variable to impute.}

\item{X_imp}{A data.frame with the fixed effects variables.}

\item{rounding_degrees}{A numeric vector with the presumed rounding degrees.}
}
\value{
A n x 1 data.frame with the original and imputed values as a factor.
}
\description{
The function uses the proportional odds logistic regression (polr) approach,
implemented in \code{mice}.
}
