% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_imp_semicont_single_2017-04-11.R
\name{imp_semicont_single}
\alias{imp_semicont_single}
\title{The function for hierarchical imputation of semicontinuous variables.}
\usage{
imp_semicont_single(y_imp, X_imp, heap = 0)
}
\arguments{
\item{y_imp}{A Vector with the variable to impute.}

\item{X_imp}{A data.frame with the fixed effects variables.}

\item{heap}{A scalar saying to which value the data might be heaped.}
}
\value{
A n x 1 data.frame with the original and imputed values.
}
\description{
The function is called by the wrapper. We consider data to be "semicontinuous" when
more than 5\% of the (non categorical) observations.\cr
For example in surveys a certain portion of people, when asked for their income,
report "0", which clearly violates the assumption of income to be (log-) normally distributed.
}
