% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_imp_semicont_single.R
\name{imp_semicont_single}
\alias{imp_semicont_single}
\title{The function for hierarchical imputation of semicontinuous variables.}
\usage{
imp_semicont_single(y_imp, X_imp, spike = NULL, pvalue = 0.2, k = Inf)
}
\arguments{
\item{y_imp}{A Vector with the variable to impute.}

\item{X_imp}{A data.frame with the fixed effects variables.}

\item{spike}{A numeric value saying to which value Y might be spiked.}

\item{pvalue}{A numeric between 0 and 1 denoting the threshold of p-values a variable in the imputation
model should not exceed. If they do, they are excluded from the imputation model.}

\item{k}{An integer defining the allowed maximum of levels in a factor covariate.}
}
\value{
A n x 1 data.frame with the original and imputed values.
}
\description{
The function is called by the wrapper. We consider data to be "semicontinuous" when
more than 5\% of the (non categorical) observations.\cr
For example in surveys a certain portion of people, when asked for their income,
report "0", which clearly violates the assumption of income to be (log-) normally distributed.
}
