% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2017-12-28.R
\name{interval2idf}
\alias{interval2idf}
\title{Transform interval variables to an interval data frame}
\usage{
interval2idf(interval)
}
\arguments{
\item{interval}{an \code{interval}}
}
\value{
an interval data frame (idf-object) with one variable (having a lower and an upper bound).
}
\description{
This function is the path from this hmi package to the linLIR package (Wiencierz, 2012).
}
