% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-01-31.R
\name{list_of_types_maker}
\alias{list_of_types_maker}
\title{Helps the user to make a list of types.}
\usage{
list_of_types_maker(data, rounding_degrees = NULL)
}
\arguments{
\item{data}{the data.frame also passed to \code{hmi}.}

\item{rounding_degrees}{A numeric vector with the presumed rounding degrees.}
}
\value{
a list with suggested types. Each list element has the name of a variable
in the data.frame. The elements contain a single character denoting the type of the variable.
See \code{get_type} for details about the variable types.
}
\description{
In \code{hmi} the user can add a list of types. This function gives him a framework
with suggestions. Of course the user can make changes by herself/himself afterwards.
For example, if a continuous variable as only two observations left, then get_type
interpret this as a binary variable and not a continuous.
}
