% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions.R
\name{sample_imp}
\alias{sample_imp}
\title{Sample imputation.}
\usage{
sample_imp(variable)
}
\arguments{
\item{variable}{A vector of size \code{n} with missing values.}
}
\value{
A n x 1 data.frame with the observed and imputed data
}
\description{
Function to sample values in a variable from other (observed) values in this variable.
So this imputation does not use further covariates.
}
\examples{
set.seed(123)
sample_imp(c(1, NA, 3, NA, 5))
}
