% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions.R
\name{suggest_rounding_degrees}
\alias{suggest_rounding_degrees}
\title{suggesting rounding degrees}
\usage{
suggest_rounding_degrees(x)
}
\arguments{
\item{x}{A vector or \code{interval} object.}
}
\description{
A function that suggests some rounding degrees of a continuous variable
(classically formatted or as interval object).
The basic idea is 1. to count which factor is observed in the data more often than expected.
2. to check whether a factor can explain at least three observed piles of observations in the data
3. to check whether a factor explains at least 20 \% of observations (additional to previous factors).
Factors fulfilling this premises are returned as suggested rounding degrees.
}
