% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmi_smallfunctions_2018-05-31.R
\name{table}
\alias{table}
\alias{table.interval}
\alias{table.default}
\title{Tabulating interval objects}
\usage{
table(x, ...)

\method{table}{interval}(x, sort = "lowerbound_increasing", ...)

\method{table}{default}(x, ...)
}
\arguments{
\item{x}{In its most save way, \code{x} is an object from class \code{interval}.}

\item{...}{Other parameters passed to \code{table}.}

\item{sort}{A character specifying how the values should be sorted if only one variable is to be plotted.
\code{sort = "lowerbound_increasing"} (the default) sorts the data primarily by their lower bound, and secondarily
(this means for equal lower bounds) by their upper bounds. Both in increasing order.
For \code{sort = "lowerbound_decreasing"} both happens in decreasing order.
\code{sort = "mostprecise_increasing"} sorts the data by their length of the interval they represent,
and within equal lengths by the lower bound. Both in increasing order.
\code{sort = "mostprecise_decreasing"} both happens in decreasing order.}
}
\value{
A table.
}
\description{
Function to tabulate interval objects
}
